swagger controladores

+++ clinicaOdontologica/src/main/java/com/backend/clinica/odontologica/dto/salida/odontologo/OdontologoSalidaDto.java
@@ -2,7 +2,7 @@ package com.backend.clinica.odontologica.dto.salida.odontologo;
 
 public class OdontologoSalidaDto {
 
-    private Long id;
+    private int id;
     private String matricula;
     private String nombre;
     private String apellido;
@@ -10,18 +10,18 @@ public class OdontologoSalidaDto {
     public OdontologoSalidaDto() {
     }
 
-    public OdontologoSalidaDto(Long id, String matricula, String nombre, String apellido) {
+    public OdontologoSalidaDto(int id, String matricula, String nombre, String apellido) {
         this.id = id;
         this.matricula = matricula;
         this.nombre = nombre;
         this.apellido = apellido;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
